FxEvent Distribution
Created bus to send events around
created game client to decode the payload once
added everyone else as subscribers to client

+++ src-tauri/src/game/frontend/fxevent.rs
@@ -1,7 +1,7 @@
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
 pub enum FxEvent {
     BallWallHit { pos: (f32, f32), intensity: f32 },
-    GoalScored { team_id: u32 },
+    GoalScored { team_id: u8 },
     // PlaySound {
     //     kind: SoundKind,
     //     pos: (f32, f32),
